pyproject.toml
[project]
name = "technical-interview"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "awscli>=1.37.0",
    "boto3>=1.36.0",
    "matplotlib>=3.10.0",
    "mlflow>=2.19.0",
    "pandas>=2.2.3",
    "pyarrow>=18.1.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "scikit-learn>=1.6.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
